pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cipipe"
version = "0.1.0"
description = "CIPipeline is a tool to analyze and process Calcium Imaging data."
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Gonzalez Agustín", email = "agngonzalez@fi.uba.ar" },
    { name = "Loyarte Iván", email = "iloyarte@fi.uba.ar" },
    { name = "Rueda Nazarena", email = "nrueda@fi.uba.ar" },
    { name = "Singer Joaquín", email = "josinger@fi.uba.arv" }
]
dependencies = [
    "rich>=13.7.0",
    "PyYAML>=6.0"
]

[project.urls]
Homepage = "https://github.com/CGK-Laboratory/ci_pipe"
Repository = "https://github.com/CGK-Laboratory/ci_pipe"

[tool.setuptools.packages.find]
where = ["src"]